[data 활용] Assignment 2 미션 리팩토링중
