+++ public/data/HyukskeeFeed.json
@@ -1,0 +1,86 @@
+[
+  {
+    "userId": "HYUKSKEE",
+    "ID": 1,
+    "HeaderImg": "images/JHKim/운명 트로트 2.jpg",
+    "TitleImg": "images/JHKim/야나두.gif"
+  },
+  {
+    "userId": "KIMSKEE",
+    "ID": 2,
+    "HeaderImg": "images/JHKim/0.jpg",
+    "TitleImg": "images/JHKim/tag.png"
+  },
+  {
+    "userId": "JINSKEE",
+    "ID": 3,
+    "HeaderImg": "images/JHKim/1.png",
+    "TitleImg": "images/JHKim/heart.png"
+  },
+  {
+    "userId": "PARKSKEE",
+    "ID": 4,
+    "HeaderImg": "images/JHKim/2.JPG",
+    "TitleImg": "images/JHKim/export.png"
+  },
+  {
+    "userId": "CHOISKEE",
+    "ID": 5,
+    "HeaderImg": "images/JHKim/3.png",
+    "TitleImg": "images/JHKim/chat.png"
+  },
+  {
+    "userId": "LEESKEE",
+    "ID": 6,
+    "HeaderImg": "images/JHKim/4.jfif",
+    "TitleImg": "images/JHKim/운명 트로트 2.jpg"
+  },
+  {
+    "userId": "JUNGSKEE",
+    "ID": 7,
+    "HeaderImg": "images/JHKim/5.png",
+    "TitleImg": "images/JHKim/7.png"
+  },
+  {
+    "userId": "YANGKSEE",
+    "ID": 8,
+    "HeaderImg": "images/JHKim/6.jpg",
+    "TitleImg": "images/JHKim/5.png"
+  },
+  {
+    "userId": "JOSKEE",
+    "ID": 9,
+    "HeaderImg": "images/JHKim/7.png",
+    "TitleImg": "images/JHKim/6.jpg"
+  },
+  {
+    "userId": "JANGSKEE",
+    "ID": 10,
+    "HeaderImg": "images/JHKim/야나두.gif",
+    "TitleImg": "images/JHKim/4.jfif"
+  },
+  {
+    "userId": "WANGSKEE",
+    "ID": 11,
+    "HeaderImg": "images/JHKim/chat.png",
+    "TitleImg": "images/JHKim/0.jpg"
+  },
+  {
+    "userId": "YOOSKEE",
+    "ID": 12,
+    "HeaderImg": "images/JHKim/export.png",
+    "TitleImg": "images/JHKim/1.png"
+  },
+  {
+    "userId": "HAKSKEE",
+    "ID": 13,
+    "HeaderImg": "images/JHKim/heart.png",
+    "TitleImg": "images/JHKim/2.JPG"
+  },
+  {
+    "userId": "GOSKEE",
+    "ID": 14,
+    "HeaderImg": "images/JHKim/tag.png",
+    "TitleImg": "images/JHKim/3.png"
+  }
+]
